requirements.txt
# Backend
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
asyncpg>=0.29.0

# Auth & Sessions
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
itsdangerous>=2.1.2

# Telegram
aiogram>=3.3.0
httpx>=0.26.0

# Templates
jinja2>=3.1.0

# Optional: Redis
redis>=5.0.0

# Scheduler
apscheduler>=3.10.0

# Utils
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
